// src/model/UIElements.ts
export interface UIElements {
  pageNumberBtn: HTMLElement | undefined;
  zoomPanel: HTMLDivElement | undefined;
  buttonsPanel: HTMLDivElement | undefined;
  fullScreenModeButton: HTMLElement | undefined;
  increaseFontButton: HTMLElement | undefined;
  decreaseFontButton: HTMLElement | undefined;
  increaseLineHeight: HTMLElement | undefined;
  decreaseLineHeight: HTMLElement | undefined;
  nightModeButton: HTMLElement | undefined;
  sepiaModeButton: HTMLElement | undefined;
  verticalScrollButton: HTMLElement | undefined;
}